+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `fast-assoc-in` - uses `fast-assoc` and `val-at` directly
 - `fast-updatge-in` - same as `fast-assoc-in`.
+- `definline+` - like `definline` but takes multiple arities
 
 ### Benchmarks